Hop 11 add hopt flseq (#8)
* Add HOPT_FLSEQ and HOPT_FLFLOW

OPT_FLAG_SEQUENCE allows you to redefine an option (with overwrite or no if HOPT_FLOW is bits set, no compatible with HOPT_FLCB). Useful with HOPT_FLCB to repeat a callback function.

* Add HOPT_FLSEQ and HOPT_FLFLOW

OPT_FLAG_SEQUENCE allows you to redefine an option (with overwrite or no if HOPT_FLOW is bits set, no compatible with HOPT_FLCB). Useful with HOPT_FLCB to repeat a callback function.